% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_tb.R
\name{$.tb}
\alias{$.tb}
\title{Subset `tb` objects or pipe `[.tb` calls}
\usage{
\method{$}{tb}(e1, e2)
}
\description{
`some_tb$.` simply returns `some_tb`, which makes piping easy. Piping with
`magrittr`'s `%>%` operator is possible as well but requires another package,
is less efficient, more verbose, and has less convenient precedence.
}
