% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_tb_select.R
\name{s}
\alias{s}
\title{select and transform columns}
\usage{
s(...)
}
\arguments{
\item{...}{expressions}
}
\description{
This reproduces the functionality of SQL select, `dplyr::transmute()` or
*data.table*'s use of `.()` in `j`.
}
