% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_tb.R
\name{tb}
\alias{tb}
\title{Create tb object}
\usage{
tb(...)
}
\description{
`tb()` works a lot like `data.frame()`, but allows for `list` or `data.frame`
columns, and outputs a data frame with the additional class `tb`
}
