% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_bracket.R
\name{[.tb}
\alias{[.tb}
\title{Modify a tb object}
\usage{
\method{[}{tb}(x, i, j, ..., by, .stack = NULL, fill = NULL, drop = FALSE)
}
\arguments{
\item{x}{tb object}

\item{i}{numeric, logical, character or formula to subset rows, if a formula
the lhs must evaluate to numeric and the rhs specifies the variables to
slice along.}

\item{j}{numeric, logical or character to subset columns}

\item{...}{Name-value pairs of expression to be evaluated in the context of
the tb, see details}

\item{by}{variables to aggregate by}

\item{.stack}{to pivot to longer, see `?tb_stack`}

\item{drop}{not used}
}
\description{
`...` can contain `foo = expr` arguments such as those
}
