% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_bracket.R
\name{[.tb}
\alias{[.tb}
\title{Modify a tb object}
\usage{
\method{[}{tb}(x, i, j, ..., by, drop = FALSE, fill = NULL)
}
\arguments{
\item{x}{teebee object}

\item{i}{numeric, logical, character or formula to subset rows, i a formula
the lhs must evaluate to numeric and the rhs specifies the variables to
slice along.}

\item{j}{numeric, logical or character to subset columns}

\item{...}{Name-value pairs of expression to be evaluated in the context of
the teebee, see details}

\item{by}{variables to aggregate by}

\item{.rm}{if TRUE columns not created by the call or part of by cols will be
removed}

\item{.unchop}{}
}
\value{

}
\description{
`...` can contain `foo = expr` arguments such as those
}
